Ensure `_channel` is always lexically defined
If there's an exception thrown as we're calling `SFTPClient.from_transport()`,
then `_channel` won't ever get created in the lexical environment and this
causes a `UnboundLocalError` on top of the exception that's getting passed
along, creating unnecessary noise. Fix this by always defining the variable, and
then making sure it is non-`None` before trying to close it.